+++ src/controllers/index.js
@@ -32,6 +32,7 @@ Controllers.mods = require('./mods');
 Controllers.sitemap = require('./sitemap');
 Controllers['404'] = require('./404');
 Controllers.errors = require('./errors');
+Controllers.handleOSDRequest = require('./osd');
 
 Controllers.home = function (req, res, next) {
 	var route = meta.config.homePageRoute || (meta.config.homePageCustom || '').replace(/^\/+/, '') || 'categories';
