+++ src/routes/meta.js
@@ -8,5 +8,5 @@ module.exports = function (app, middleware, controllers) {
 	app.get('/robots.txt', controllers.robots);
 	app.get('/manifest.json', controllers.manifest);
 	app.get('/css/previews/:theme', controllers.admin.themes.get);
-	app.get('/osd.xml', require('../meta').osd.handleOSDRequest);
+	app.get('/osd.xml', controllers.handleOSDRequest);
 };
